+++ modules/test-cpp/src/main.cpp
@@ -8,6 +8,16 @@ bool test_cpp::test_2_args(
   return true;
 }
 
+std::optional<std::uint32_t> test_cpp::test_optional(
+  const std::optional<std::uint32_t> &a
+)
+{
+  if (a) {
+    return *a + 1;
+  }
+  return std::nullopt;
+}
+
 int main(int argc, char *argv[])
 {
 }